fix: 프로필 사진이 없는 경우 NPE 가 발생하는 문제 수정
fix: 프로필 사진이 없는 경우 NPE 가 발생하는 문제 수정

+++ myStory/src/main/java/com/team/mystory/account/user/service/LoginService.java
@@ -117,7 +117,10 @@ public class LoginService {
 	public String modifyProfileImage(String accessToken, MultipartFile multipartFile) throws AccountException, IOException {
 		User result = findByUserByAccessToken(accessToken);
 
-		s3Service.deleteFile(result.getProfileImage());
+		if(result.getProfileImage() != null && !result.getProfileImage().isEmpty()) {
+			s3Service.deleteFile(result.getProfileImage());
+		}
+
 		String url = s3Service.uploadImageToS3(multipartFile);
 
 		result.updateProfileImage(url);
